requirements.txt
python-dateutil~=2.8.2
requests~=2.31.0
Pillow~=10.1.0
StringProgressBar~=1.1.1
python-dotenv~=1.0.0
Flask~=3.0.1
numpy~=1.26.2
schedule~=1.2.1
waitress~=2.1.2
py-cord~=2.5.0
aiohttp>=3.8.0
psycopg2-binary~=2.9.10
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-api-python-client>=2.0.0
emoji>=2.2.0
Flask-Cors~=5.0.1
openai>=1.75.0
PyNaCl>=1.5.0
boto3>=1.34.0
certifi>=2024.1.0